+++ src/Token/SecretValidator.php
@@ -7,11 +7,11 @@ use ReallySimpleJWT\Interfaces\Secret as SecretInterface;
 class SecretValidator implements SecretInterface
 {
     /**
-     * Simplified validation to 8+ characters, must have a digit and upper or lower case character
+     * Simplified validation to 8+ characters, must have upper or lower case character
      * @inheritDoc
      */
     public function validate(string $secret): bool
     {
-        return (bool)preg_match('/^.*(?=.{8,}+)(?=.*\d+)(?=.*[A-Za-z]+).*$/', $secret);
+        return (bool)preg_match('/^.*(?=.{8,}+)(?=.*[A-Za-z]+).*$/', $secret);
     }
 }