+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3.0
+
+- Added ability to extend expired sessions when default Storefront API calls are made
+- Changed validation to accept secret without numbers
+
 # 0.2.0
 
 - Added `close` button to success purchase page